création compte et ajout sauce  fonctionnels

+++ controllers/sauce.js
@@ -1,7 +1,10 @@
 const Sauce = require('../models/sauce');
+const fs = require('fs');
 
-exports.createSauce = (req, res, next) => {
+exports.createSauce = (req, res) => {
   const sauceObject = JSON.parse(req.body.sauce);
+  console.log(sauceObject);
+
   delete sauceObject._id;
   delete sauceObject._userId;
 
@@ -32,45 +35,65 @@ exports.getOneSauce = (req, res, next) => {
 };
 
 exports.modifySauce = (req, res, next) => {
-    Sauce.updateOne({_id: req.params.id}, {...req.body, _id: req.params.id})
-    .then(() => {
-      res.status(201).json({
-         message: 'Commentaire modifié!'
-      });
-    })
-    .catch(
-      (error) => {
-        res.status(400).json({
-          error: error
+  const sauceObject = req.file ? {
+    ... JSON.parse(req.body.sauce),
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+
+  }: {...req.body};
+
+  delete sauceObject._userId;
+  Sauce.findOne({_id: req.auth.id})
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId){
+        return res.status(401).json({ message: 'Action non autorisé'});
+      } 
+
+      Sauce.updateOne({_id: req.params.id}, {...sauceObject, _id: req.params.id})
+      .then(()  =>
+        res.status(200).json( { message: 'Sauce modifié'} )
+      )
+      .catch(error => res.status(401).json({error}));
+      
+    })    
+    .catch((error) =>{
+      res.status(400).json({error})
+    });  
+    
+};    
+ 
+  
+exports.deleteSauce = (req, res, next) => {
+  Sauce.findOne({_id: req.auth.id})
+    .then(Sauce => {
+      if (Sauce.userId != req.auth.userId){
+        res.status(401).json({ message: 'Action non autorisé'});
+      } else {
+        const filename = thing.imageUrl.split('/image/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id})
+            .then(()=> {
+              res.status(200).json({message: 'sauce supprimé'})}
+            )
+            .catch(error => 
+              res.status(401).json({ error})
+            );
         });
       }
-    );
-  };
-  
-  exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({_id: req.params.id})
-    .then(() => {
-      res.status(200).json({
-        message: 'Supprimé'
-      });
     })
-    .catch(
-      (error) => {
-        res.status(400).json({
-          error: error
-        });
-      }
-    );
-  };
+    .catch( error => {
+      res.status(500).json({error});
+    });    
+   
+};
   
-  exports.getAllSauce = (req, res, next) => {
-    Sauce.find()
-    .then((sauces) => {
-      res.status(200).json(Sauces);
+exports.getAllSauce = (req, res, next) => {
+  Sauce.find()
+    .then((Sauce) => {
+      res.status(200).json(Sauce);
     })
     .catch((error) => {
       res.status(400).json({
         error: error
       });
     });
-  };
+};
